+++ packages/paste-core/components/button/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.22](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@2.3.21...@twilio-paste/button@2.3.22) (2020-10-22)
+
+**Note:** Version bump only for package @twilio-paste/button
+
+
+
+
+
+## [2.3.21](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@2.3.20...@twilio-paste/button@2.3.21) (2020-10-21)
+
+**Note:** Version bump only for package @twilio-paste/button
+
+
+
+
+
+## [2.3.20](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@2.3.19...@twilio-paste/button@2.3.20) (2020-10-19)
+
+**Note:** Version bump only for package @twilio-paste/button
+
+
+
+
+
 ## [2.3.19](https://github.com/twilio-labs/paste/compare/@twilio-paste/button@2.3.18...@twilio-paste/button@2.3.19) (2020-10-15)
 
 **Note:** Version bump only for package @twilio-paste/button